Added in the first draft of patches
Patches can now simply be created using a bootstrapped version of gresh. The
command for this is as follows:

$ gresh patch git@my-remote.com:user/repo.git branch-name

This will generate the git request-pull and ask if you're sure that you'd like
to submit the patch.

Please note this is a first draft of this code and has not been fully tested.